tests: Move NoRecipientIDsTest to test_subs.py.
